+++ service/parsers/erc1155_batch.go
@@ -52,8 +52,8 @@ func ERC1155Batch(log types.Log) ([]*events.Transfer, error) {
 			TransactionHash:   log.TxHash.Hex(),
 			SenderAddress:     log.Topics[2].Hex(),
 			ReceiverAddress:   log.Topics[3].Hex(),
-			TokenCount:        count.Uint64(),
-			// EmmittedAt set after parsing
+			TokenCount:        uint(count.Uint64()),
+			// EmittedAt set after parsing
 		}
 		transfers = append(transfers, &transfer)
 	}
