+++ service/parsers/erc1155_transfer.go
@@ -47,8 +47,8 @@ func ERC1155Transfer(log types.Log) (*events.Transfer, error) {
 		TransactionHash:   log.TxHash.Hex(),
 		SenderAddress:     log.Topics[2].Hex(),
 		ReceiverAddress:   log.Topics[3].Hex(),
-		TokenCount:        count.Uint64(),
-		// EmmittedAt set after parsing
+		TokenCount:        uint(count.Uint64()),
+		// EmittedAt set after parsing
 	}
 
 	return &transfer, nil
